+++ src/app/layout.tsx
@@ -2,10 +2,9 @@ import type { Metadata, Viewport } from 'next'
 import { Inter } from 'next/font/google'
 import './globals.css'
 import 'react-toastify/dist/ReactToastify.css'
-import Script from 'next/script'
 import { NuqsAdapter } from 'nuqs/adapters/next/app'
 import { ToastContainer } from 'react-toastify'
-import { umamiWebsiteId } from '@/lib/utils/umami'
+import { UmamiScript } from '@/components/home/UmamiScript'
 
 const inter = Inter({ subsets: ['latin'] })
 
@@ -27,11 +26,7 @@ export default function RootLayout({ children }: { children: React.ReactNode }) 
     return (
         <html lang="en">
             <head>
-                <Script
-                    src="https://umami-cmf.vercel.app/script.js"
-                    data-website-id={umamiWebsiteId}
-                    strategy="afterInteractive"
-                />
+                <UmamiScript />
             </head>
             <body className={inter.className}>
                 <NuqsAdapter>{children}</NuqsAdapter>